Media Placeholder: make component easier to read and reduce required customization
